Updated to VisUI 1.1.0. #23

+++ lml-vis/src/main/java/com/github/czyzby/lml/vis/parser/impl/attribute/grid/ItemHeightLmlAttribute.java
@@ -5,7 +5,7 @@ import com.github.czyzby.lml.parser.tag.LmlAttribute;
 import com.github.czyzby.lml.parser.tag.LmlTag;
 import com.kotcrab.vis.ui.layout.GridGroup;
 
-/** See {@link GridGroup#setItemHeight(int)}. Mapped to "itemHeight".
+/** See {@link GridGroup#setItemHeight(float)}. Mapped to "itemHeight".
  *
  * @author MJ */
 public class ItemHeightLmlAttribute implements LmlAttribute<GridGroup> {
@@ -17,6 +17,6 @@ public class ItemHeightLmlAttribute implements LmlAttribute<GridGroup> {
     @Override
     public void process(final LmlParser parser, final LmlTag tag, final GridGroup actor,
             final String rawAttributeData) {
-        actor.setItemHeight(parser.parseInt(rawAttributeData, actor));
+        actor.setItemHeight(parser.parseFloat(rawAttributeData, actor));
     }
 }
